java-note | Move Streams package and Lambdas package into Lambdas and Streams package
