+++ JavaCore/src/main/java/lambdas_streams/lambdas/LambdasAndFunctionalPrograming.md
