+++ JavaCore/src/main/java/lambdas_streams/streams/Source.md
