+++ JavaCore/src/main/java/lambdas_streams/streams/StreamDemo.java
@@ -1,6 +1,5 @@
-package streams;
+package lambdas_streams.streams;
 
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
