+++ JavaCore/src/main/java/lambdas_streams/streams/TerminalOperation.md
