+++ JavaCore/src/main/java/lambdas_streams/streams/optional/App.java
@@ -1,4 +1,4 @@
-package streams.optional;
+package lambdas_streams.streams.optional;
 
 import java.util.Arrays;
 import java.util.NoSuchElementException;
